Add Matrix Arrays, logic, and stop/start functions
Code should produce an array of to the screen width/13 arrays filled up till screen height/13.
--
A class was created in matrixChar
to signfy characters, and strength we can use to lower intensity of characters
--
Two functions created one addMatrix to randomly decide if it should add a Character
One moveMatrix that checks the array and removes strength 1 characters, and adds another character ahead of strength 5 character.